+++ lib/fetchJson.ts
@@ -1,0 +1,8 @@
+
+export async function fetchJson(url: RequestInfo, init?: RequestInit): Promise<any> {
+  const response = await fetch(url, init);
+  const json = await response.json();
+
+  if (!response.ok) throw new Error(`${response.statusText}: ${json.error?.message || JSON.stringify(json)}`);
+  return json;
+}
