+++ lib/utils.ts
@@ -74,13 +74,6 @@ export function getEtherscanLink(
   }
 }
 
-export async function fetchJson(url: RequestInfo, init?: RequestInit): Promise<any> {
-  const response = await fetch(url, init);
-  const json = await response.json();
-
-  if (!response.ok) throw new Error(`${response.statusText}: ${json.error?.message || JSON.stringify(json)}`);
-  return json;
-}
 
 /* eslint-disable no-useless-escape */
 // https://medium.com/@mhagemann/the-ultimate-way-to-slugify-a-url-string-in-javascript-b8e4a0d849e1
