+++ pages/address/[address]/index.tsx
@@ -7,7 +7,7 @@ import Link from 'next/link';
 import Head from 'next/head';
 import { Icon } from '@makerdao/dai-ui-icons';
 import { getNetwork } from 'lib/maker';
-import { fetchJson } from 'lib/utils';
+import { fetchJson } from 'lib/fetchJson';
 import { useAnalytics } from 'lib/client/analytics/useAnalytics';
 import { ANALYTICS_PAGES } from 'lib/client/analytics/analytics.constants';
 import PrimaryLayout from 'components/layouts/Primary';
