+++ pages/delegates/index.tsx
@@ -4,7 +4,7 @@ import { Heading, Box, Card, Text, Link as ThemeUILInk, jsx } from 'theme-ui';
 import { GetStaticProps } from 'next';
 import Head from 'next/head';
 import ErrorPage from 'next/error';
-import { isDefaultNetwork } from 'lib/maker';
+import { isDefaultNetwork, isTestnet } from 'lib/maker';
 import { fetchDelegates } from 'modules/delegates/api/fetchDelegates';
 import { DelegateStatusEnum } from 'modules/delegates/delegates.constants';
 import { shuffleArray } from 'lib/common/shuffleArray';
@@ -16,7 +16,7 @@ import ResourceBox from 'components/ResourceBox';
 import { DelegateCard } from 'modules/delegates/components';
 import PageLoadingPlaceholder from 'components/PageLoadingPlaceholder';
 import { getNetwork } from 'lib/maker';
-import { fetchJson } from 'lib/utils';
+import { fetchJson } from 'lib/fetchJson';
 import { useAnalytics } from 'lib/client/analytics/useAnalytics';
 import { ANALYTICS_PAGES } from 'lib/client/analytics/analytics.constants';
 import useAccountsStore from 'stores/accounts';
