+++ pages/polling.tsx
@@ -34,7 +34,7 @@ import PageLoadingPlaceholder from 'components/PageLoadingPlaceholder';
 import { useAnalytics } from 'lib/client/analytics/useAnalytics';
 import { ANALYTICS_PAGES } from 'lib/client/analytics/analytics.constants';
 import { getPolls } from 'modules/polling/api/fetchPolls';
-import { fetchJson } from 'lib/utils';
+import { fetchJson } from 'lib/fetchJson';
 
 type Props = {
   polls: Poll[];
